done all changes based on suggestions

+++ backend/src/services/user.services.ts
@@ -5,49 +5,40 @@ import { User } from "../Interface/user";
 
 const JWT_SECRET: string = process.env.JWT_SECRET as string;
 
-export const AllUsersData = async (): Promise<User> => await User.find();
+export const getAllUsers = async (): Promise<User> => await User.find().select("-password");
 
-export const UserDataById = async (id: string): Promise<User> =>
-  await User.findById(id);
+export const getUserById = async (id: string): Promise<User> =>
+  await User.findById(id).select("-password");
 
-export const signUpUser = async (data: User): Promise<User> => {
-  const {
-    name,
-    email,
-    password,
-    phone,
-    address,
-    country,
-    businessCategory,
-    role,
-  } = data;
-  const hashedPassword = await bcrypt.hash(password, 12);
-  const user = new User({
-    name,
-    email,
-    password: hashedPassword,
-    phone,
-    address,
-    country,
-    businessCategory,
-    role,
-  });
+export const addUser = async (data: User): Promise<User> => {
+  data.password = await bcrypt.hash(data.password, 12);
+  const user = new User(data);
   return await user.save();
 };
 
-export const signInUser = async (
+export const loginUser = async (
   email: string,
   password: string
-): Promise<User> => {
+): Promise<string> => {
   const user = await User.findOne({ email });
-  const token = jwt.sign({ userId: user._id }, JWT_SECRET, {
-    expiresIn: "1h",
-  });
-  return token;
+
+  if (user && (await bcrypt.compare(password, user.password))) {
+    return jwt.sign({ userId: user._id }, JWT_SECRET, {
+      expiresIn: "1h",
+    });
+  }
+
+  throw "Invalid email or password.";
 };
 
-export const updateUser = async (id: string, data: User): Promise<User> =>
-  await User.findByIdAndUpdate(id, data, { new: true });
+export const updateUser = async (id: string, data: User): Promise<User> => {
+  const existUser = await User.findById(id);
+  if (!existUser) throw "User is not exist";
+  return await User.findByIdAndUpdate(id, data, { new: true });
+};
 
 export const deleteUser = async (id: string): Promise<User> =>
-  await User.findByIdAndDelete(id);
+{  const existUser = await User.findById(id);
+  if (!existUser) throw "User is not exist";
+  return await User.findByIdAndDelete(id);
+}
